Move upgradestate logic to its own file
